pyproject.toml
[project]
name = "atmosledger"
version = "0.1.0"
description = "Enterprise-grade weather ingestion, aggregation, forecasting, and anomaly detection backend."
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "MIT" }
authors = [{ name = "Marcus L. Williams" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.32",
    "psycopg[binary]>=3.2.1",
    "alembic>=1.13.2",
    "redis>=5.0.8",
    "rq>=1.16.2",
    "structlog>=24.2.0",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
ruff = "^0.5.6"
mypy = "^1.11.1"
types-redis = "^4.6.0.20241004"
python-dotenv = "^1.0.1"


[tool.poetry]
package-mode = false
[tool.poetry.scripts]
dev = "atmosledger.util.time:noop"


[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
mypy_path = "src"
